+++ src/components/shop-inventory/index.js
@@ -4,7 +4,7 @@ import { connect }         from 'react-redux';
 import ShopItem              from './shop-item';
 import shopItems             from './shop-items';
 import ViewItem              from '../view-item';
-import uuidv4                from '../../modules/uuid-v4';
+import uuidv4                from '../../utils/uuid-v4';
 import { MAX_ITEMS_UPGRADE } from '../../config/constants';
 
 import './styles.scss';
