+++ src/features/map/random-map-gen/generate-monsters.js
@@ -1,7 +1,7 @@
 import { MAP_DIMENSIONS }  from '../../../config/constants';
 import generateMonsterType from './generate-monster-type';
-import getSurroundingTiles from '../../../modules/get-surrounding-tiles';
-import arrContainArr       from '../../../modules/arr-contain-arr';
+import getSurroundingTiles from '../../../utils/get-surrounding-tiles';
+import arrContainArr       from '../../../utils/arr-contain-arr';
 // generates random monsters for a random map
 export default function generateMonsters(floorNum, map, playerPos, playerLv) {
   let availableTiles = [];
