+++ src/utils/calculate-damage.js
